Readability Improvements
Some changes to make things easier on the eyes.

I ditched the idea of storing properties in the "Editor Plugins" Entirely, and now they will be stored in its own "DeadlyLampshade/Plugin" property_list in the ProjectSettings.

As well as expanded properties for BakedLightmaps. You can even bake it directly using the new options (if you want to)